work on loader

+++ etl/index_shapes.py
@@ -363,22 +363,25 @@ def clone_and_process_repo(shapes_graph, repo_url, branch, repo_description, git
     # TODO: load each repo in a different file in the VAD
     # curl -T shapes-rdf.ttl  https://data.index.semanticscience.org/DAV/home/dba/rdf_sink/$GIT_SERVICE.ttl -u dba:password
     # https://github.com/MaastrichtU-IDS/shapes-of-you becomes: github.com/MaastrichtU-IDS/shapes-of-you.ttl
-    shapes_graph.serialize('shapes-rdf.ttl', format='turtle')
+    if (None, None, None) in shapes_graph:
+      print("This graph contains triples about Bob!")
+      shapes_graph.serialize('shapes-rdf.ttl', format='turtle')
 
-    repo_vad_url = repo_url.replace('https://', '').replace('http://', '')
-    print('loading VAD: ' + str(repo_vad_url))
+      repo_vad_url = repo_url.replace('https://', '').replace('http://', '')
+      print('loading VAD: ' + str(repo_vad_url))
 
-    # with open('shapes-rdf.ttl', 'r') as f:
-    with open(root / '../shapes-rdf.ttl', 'r') as f:
-      print(f.read())
+      # with open('shapes-rdf.ttl', 'r') as f:
+      with open(root / '../shapes-rdf.ttl', 'r') as f:
+        print(f.read())
 
-    os.system('curl -T shapes-rdf.ttl  https://data.index.semanticscience.org/DAV/home/dba/rdf_sink/' + repo_vad_url + '.ttl -u dba:' + ENDPOINT_PASSWORD)
-    # To delete:
-    # select DB.DBA.DAV_DELETE ('/DAV/home/dba/rdf_sink/gitee.ttl', 0, 'dba', 'dba');
-    try:
-      os.remove('shapes-rdf.ttl')
-    except:
-      pass
+      os.system('curl -T shapes-rdf.ttl  https://data.index.semanticscience.org/DAV/home/dba/rdf_sink/' + repo_vad_url + '.ttl -u dba:' + ENDPOINT_PASSWORD)
+      # To delete:
+      # select DB.DBA.DAV_DELETE ('/DAV/home/dba/rdf_sink/gitee.ttl', 0, 'dba', 'dba');
+      try:
+        os.remove('shapes-rdf.ttl')
+      except:
+        pass
+      
     return shapes_graph
